Admin Page with Login Screen
Needs the database information to be implemented into it and a few more functionalities, particularly in the top taskbar. Also, it needs to have the login screen use the usernames and passwords from the database instead of the if statement that is in there now (it was used for quick testing).