Almost all displays are here
Implement database so the system can display information from the database. Implement a login that uses the Usernames and Passwords from the database. Certain buttons do not have full feature capabilities. If more displays are needed, message me on discord and send me a screenshot of the entire screen from team 3's program, so that I have a reference to create a display for our project.